lambda/requirements.txt
ask-sdk-core==1.19.0
ask-sdk-model==1.82.0
openai==0.28.1
# dict-deep==4.1.2
urllib3==1.26.18
python-dotenv==0.21.1
